make utils-package renaming things
