+++ client/utils/display.py
@@ -1,0 +1,42 @@
+black = (0,0,0)
+r = (255,0,0)
+g = (0,255,0)
+y = (255,255,0)
+gray = (180, 180, 180)
+b = (0,0,255)
+w =(255,255,255)
+triangel = [g,g,g,g,g,g,g,g,
+            g,g,g,g,g,g,g,g,
+            g,g,g,w,g,g,g,g,
+            g,g,w,g,w,g,g,g,
+            g,w,g,g,g,w,g,g,
+            w,w,w,w,w,w,w,g,
+            g,g,g,g,g,g,g,g,
+            g,g,g,g,g,g,g,g]
+
+piano =    [g,g,g,g,g,g,g,g,
+            g,g,g,g,g,g,g,g,
+            gray,black,gray,black,gray,black,gray,black,
+            gray,black,gray,black,gray,black,gray,black,
+            gray,gray,gray,gray,gray,gray,gray,gray,
+            gray,gray,gray,gray,gray,gray,gray,gray,
+            g,g,g,g,g,g,g,g,
+            g,g,g,g,g,g,g,g]
+
+guitar =   [w,g,g,g,g,g,g,g,
+            g,w,g,g,g,g,g,g,
+            g,g,w,g,w,w,g,g,
+            g,g,g,w,w,w,w,g,
+            g,g,g,w,w,w,w,g,
+            g,g,g,w,w,w,w,g,
+            g,g,g,g,w,w,g,g,
+            g,g,g,g,g,g,g,g]
+
+attention =     [r,r,r,y,y,r,r,r,
+                r,r,r,y,y,r,r,r,
+                r,r,r,y,y,r,r,r,
+                r,r,r,y,y,r,r,r,
+                r,r,r,r,r,r,r,r,
+                r,r,r,y,y,r,r,r,
+                r,r,r,y,y,r,r,r,
+                r,r,r,r,r,r,r,r]
