+++ helpers/musician_setup.sh
