+++ helpers/musician_startup.sh
